\name{dIPBH}
\alias{dIPBH}
\docType{package}
\title{
Density Function for the Inverse Power Burr-Hatke Distribution
}
\description{
Density Function for the Inverse Power Burr-Hatke Distribution
}
\arguments{
  \item{x}{a vector of variates}
  \item{alpha}{a vector of alpha values}
  \item{eta}{a vector of eta values}
  \item{log_transform}{a boolean for logarithmic calculation}
}
\usage{
dIPBH(x, alpha, eta, log_transform)	
}
\examples{
\dontrun{
dIPBH(x, alpha, eta, log_transform)
}
}
