\name{IPBH-package}
\alias{IPBH-package}
\alias{IPBH}
\docType{package}
\title{
  Inverse Power Burr-Hatke Distribution Functions
}
\description{
  Density function, distribution function, quantile function,
    and random variate generation for the Inverse Power Burr-Hatke distribution.
}
\details{
  Used the same as functions for other distributions, such as rnorm, dnorm, 
  pnorm, and qnorm.
}
\author{
Ethan Murra

Maintainer: Ethan Murra
}
\references{
  Package based on the following paper.
  
  Ahmed Z. Afify, Hassan M. Aljohani, Abdulaziz S. Alghamdi, Ahmed M. Gemeay,
  Abdullah M. Sarg, "A New Two-Parameter Burr-Hatke Distribution: Properties
  and Bayesian and Non-Bayesian Inference with Applications",
  Journal of Mathematics, vol. 2021, Article ID 1061083,
  16 pages, 2021. https://doi.org/10.1155/2021/1061083

}
\keyword{ package }
\seealso{
  dIPBH
  pIPBH
  qIPBH
  rIPBH
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
