\name{pIPBH}
\alias{pIPBH}
\docType{package}
\title{
Distribution Function for the Inverse Power Burr-Hatke Distribution
}
\description{
Distribution Function for the Inverse Power Burr-Hatke Distribution
}
\arguments{
  \item{x}{a vector of variates}
  \item{alpha}{a vector of alpha values}
  \item{eta}{a vector of eta values}
  \item{lower_tail}{a boolean for lower tail calculation}
  \item{log_p}{a boolean for logarithmic calculation}
}
\usage{
pIPBH(x, alpha, eta, lower_tail, log_p)	
}
\examples{
\dontrun{
pIPBH(x, alpha, eta, lower_tail, log_p)
}
}
