\name{qIPBH}
\alias{qIPBH}
\docType{package}
\title{
Quantile Function for the Inverse Power Burr-Hatke Distribution
}
\description{
Quantile Function for the Inverse Power Burr-Hatke Distribution
}
\arguments{
  \item{p}{a vector of probabilities}
  \item{alpha}{a vector of alpha values}
  \item{eta}{a vector of eta values}
  \item{lower_tail}{a boolean for lower tail calculation}
  \item{log_p}{a boolean for logarithmic calculation}
}
\usage{
qIPBH(p, alpha, eta, lower_tail, log_p)	
}
\examples{
\dontrun{
qIPBH(p, alpha, eta, lower_tail, log_p)
}
}
