\name{rIPBH}
\alias{rIPBH}
\docType{package}
\title{
Random Variate Generation for the Inverse Power Burr-Hatke Distribution
}
\description{
Random Variate Generation for the Inverse Power Burr-Hatke Distribution
}
\arguments{
  \item{n}{an integer for the number of random variates to be generated}
  \item{alpha}{a vector of alpha values}
  \item{eta}{a vector of eta values}
}
\usage{
rIPBH(n, alpha, eta)	
}
\examples{
\dontrun{
rIPBH(n, alpha, eta)
}
}
